// x-pack/platform/plugins/shared/agent_builder/server/services/tools/tool_provider.d.ts
import type { MaybePromise } from '@kbn/utility-types';
import type { KibanaRequest } from '@kbn/core-http-server';
import type { ToolType } from '@kbn/agent-builder-common';
import type { InternalToolDefinition, ToolCreateParams, ToolUpdateParams } from '@kbn/agent-builder-server/tools';
export type ToolTypeCreateParams<TConfig extends object = {}> = ToolCreateParams<TConfig>;
export type ToolTypeUpdateParams<TConfig extends object = {}> = ToolUpdateParams<TConfig>;
export interface ToolProviderListFilters {
    types?: ToolType[];
    tags?: string[];
}
export interface ReadonlyToolProvider {
    id: string;
    readonly: true;
    has(toolId: string): MaybePromise<boolean>;
    get(toolId: string): MaybePromise<InternalToolDefinition>;
    list(filters?: ToolProviderListFilters): MaybePromise<Array<InternalToolDefinition>>;
}
export interface WritableToolProvider extends Omit<ReadonlyToolProvider, 'readonly'> {
    readonly: false;
    create(params: ToolTypeCreateParams): MaybePromise<InternalToolDefinition>;
    update(toolId: string, update: ToolTypeUpdateParams): MaybePromise<InternalToolDefinition>;
    delete(toolId: string): MaybePromise<boolean>;
}
export type ToolProviderFn<Readonly extends boolean> = (opts: {
    request: KibanaRequest;
    space: string;
}) => MaybePromise<Readonly extends true ? ReadonlyToolProvider : WritableToolProvider>;
export type ToolProvider = ReadonlyToolProvider | WritableToolProvider;
export declare const isReadonlyToolProvider: (provider: ToolProvider) => provider is ReadonlyToolProvider;
export declare const isWritableToolProvider: (provider: ToolProvider) => provider is WritableToolProvider;